fix: looks complete ready for pr
